Update index.tsx
Allow for hitSlop to be set for the PanGestureHandler, this allows to specify "drag handles" on the side while still allowing for ScrollViev scrolling when the DraggableFlatlist is nested within a ScrollView. Also allows having a "activation toggle" but setting the hitSlop to full screen width or 0 through state variable.